fix(AIAI-575): move governance history classes out of internal packages
