+++ smart-workflow-dashboard/src/com/axonivy/utils/smart/workflow/governance/ui/entity/CaseHistoryGroup.java
@@ -4,7 +4,7 @@ import java.util.Comparator;
 import java.util.List;
 
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry;
-import com.axonivy.utils.smart.workflow.governance.history.internal.CaseService;
+import com.axonivy.utils.smart.workflow.governance.history.service.CaseService;
 
 import ch.ivyteam.ivy.workflow.ICase;
