+++ smart-workflow-test/pom.xml
@@ -68,6 +68,22 @@
     </dependency>
     <dependency>
       <groupId>com.axonivy.utils.ai</groupId>
+      <artifactId>smart-workflow-dashboard</artifactId>
+      <version>${project.version}</version>
+      <type>iar</type>
+      <exclusions>
+        <exclusion>
+          <groupId>*</groupId>
+          <artifactId>asm</artifactId>
+        </exclusion>
+        <exclusion>
+          <groupId>*</groupId>
+          <artifactId>jspecify</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+    <dependency>
+      <groupId>com.axonivy.utils.ai</groupId>
       <artifactId>smart-workflow-ollama</artifactId>
       <version>${project.version}</version>
       <type>iar</type>
