+++ smart-workflow-test/src_test/com/axonivy/utils/smart/workflow/governance/history/TestChatHistoryJsonParser.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry;
-import com.axonivy.utils.smart.workflow.governance.history.internal.ChatHistoryJsonParser;
 
 import ch.ivyteam.ivy.environment.IvyTest;
