+++ smart-workflow/src/com/axonivy/utils/smart/workflow/governance/ui/entity/AgentConversationView.java
@@ -6,8 +6,8 @@ import java.util.List;
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry;
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry.GuardrailExecution;
 import com.axonivy.utils.smart.workflow.governance.history.entity.AgentConversationEntry.ToolExecution;
-import com.axonivy.utils.smart.workflow.governance.history.internal.ChatHistoryJsonParser;
-import com.axonivy.utils.smart.workflow.governance.history.internal.ChatHistoryJsonParser.TokenUsage;
+import com.axonivy.utils.smart.workflow.governance.history.ChatHistoryJsonParser;
+import com.axonivy.utils.smart.workflow.governance.history.ChatHistoryJsonParser.TokenUsage;
 import com.axonivy.utils.smart.workflow.governance.utils.DatePatternUtils;
 
 public class AgentConversationView {
